Refactor NativeMethods
- move file to subdirectory
- split and move function pointers to child classes that get
  dynamically allocated. This should improve the memory footprint
  because it only allocates when methods are accessed in an area.
